moved classes to their right place
