+++ src/main/java/org/rulez/demokracia/pdengine/DBSessionManager.java
