+++ src/test/java/org/rulez/demokracia/pdengine/testhelpers/VoteInvariantCheck.java
@@ -1,11 +1,10 @@
-package org.rulez.demokracia.pdengine;
+package org.rulez.demokracia.pdengine.testhelpers;
 
 import static org.junit.Assert.assertEquals;
 
 import java.util.ArrayList;
 import org.junit.After;
 import org.rulez.demokracia.pdengine.dataobjects.VoteEntity;
-import org.rulez.demokracia.pdengine.testhelpers.ThrowableTester;
 
 public class VoteInvariantCheck extends ThrowableTester {
